Move CuraEngine profile logic from profiles.py into cura.py (#325)
Extract all CuraEngine-specific code from profiles.py into cura.py:
- load_cura_definition_map(), _deep_merge_cura_overrides()
- extract_cura_docker_defs(), _squash_cura_def()
- pin_cura_definitions()

profiles.py now contains no CuraEngine-specific logic — it re-exports
for backward compatibility. 653 tests pass, zero lint/type errors.
