+++ changes/+move-cura-profiles.misc
@@ -1,0 +1,1 @@
+Move CuraEngine definition pinning and profile logic from ``profiles.py`` into ``cura.py`` engine module.
